Extract coordinator refresh runner
